+++ core/interfaces.go
@@ -34,6 +34,7 @@ type Services interface {
 	IsGroupAdmin(clientID string, groupID string, userID string) (bool, error)
 
 	CreateGroup(clientID string, current *model.User, group *model.Group, membersConfig *model.DefaultMembershipConfig) (*string, *utils.GroupError)
+	CreateGroupV3(clientID string, current *model.User, group *model.Group, membershipStatuses model.MembershipStatuses) (*string, *utils.GroupError)
 	UpdateGroup(clientID string, current *model.User, group *model.Group) *utils.GroupError
 	UpdateGroupDateUpdated(clientID string, groupID string) error
 	DeleteGroup(clientID string, current *model.User, id string) error
@@ -141,6 +142,10 @@ func (s *servicesImpl) IsGroupAdmin(clientID string, groupID string, userID stri
 
 func (s *servicesImpl) CreateGroup(clientID string, current *model.User, group *model.Group, membersConfig *model.DefaultMembershipConfig) (*string, *utils.GroupError) {
 	return s.app.createGroup(clientID, current, group, membersConfig)
+}
+
+func (s *servicesImpl) CreateGroupV3(clientID string, current *model.User, group *model.Group, membershipStatuses model.MembershipStatuses) (*string, *utils.GroupError) {
+	return s.app.createGroupV3(clientID, current, group, membershipStatuses)
 }
 
 func (s *servicesImpl) UpdateGroup(clientID string, current *model.User, group *model.Group) *utils.GroupError {