+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Changed
+- Improve PUT group/{group-id}/members/multi-create [#562](https://github.com/rokwire/groups-building-block/issues/562)
+### Added
+- Improve POST api/groups [#561](https://github.com/rokwire/groups-building-block/issues/561)
+
 ## [1.62.0] - 2025-04-17
 ### Changed
 - Support Google Trust Services as CA [#558](https://github.com/rokwire/groups-building-block/issues/558)